Split out tests for each language
